refactor progress component
